+++ frontend/src/domain/use_cases/missions/createOrEditMission.ts
@@ -7,7 +7,7 @@ export const createOrEditMissionAndGoToMissionsList = values => dispatch => {
   const upsertMission = !values.id ? missionsAPI.endpoints.createMission : missionsAPI.endpoints.updateMission
   dispatch(upsertMission.initiate(values)).then(response => {
     if ('data' in response) {
-      dispatch(sideWindowActions.openAndGoTo(sideWindowPaths.MISSIONS))
+      dispatch(sideWindowActions.focusAndGoTo(sideWindowPaths.MISSIONS))
     } else {
       dispatch(setError(response.error))
     }
